cherry pick light/default rename
