+++ packages/react-theme/src/themes/web/lightTheme.ts
@@ -1,6 +1,6 @@
-import { createNeutralColorTokens, sharedColorTokens } from '../../default';
-import { createShadowLevelTokens } from '../../default';
-import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights } from '../../global';
+import { createNeutralColorTokens, sharedColorTokens } from '../../utils/light/index';
+import { createShadowLevelTokens } from '../../utils/light/index';
+import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights } from '../../global/index';
 import { Theme } from '../../types';
 
 const common = {
@@ -14,7 +14,7 @@ const common = {
 
 const webNeutralColorTokens = createNeutralColorTokens(brandColors.web);
 
-export const webDefaultTheme: Theme = {
+export const webLightTheme: Theme = {
   brandColors: brandColors.web,
   neutralColorTokens: webNeutralColorTokens,
   shadowLevels: createShadowLevelTokens(
