+++ packages/react-theme/src/utils/light/shadows.ts
@@ -1,4 +1,4 @@
-import { ShadowLevelTokens } from '../types';
+import { ShadowLevelTokens } from '../../types';
 
 export function createShadowLevelTokens(ambientColor: string, keyColor: string): ShadowLevelTokens {
   return {
